requirements-dev.txt
# Development Dependencies
# Optional tools for enhanced development experience

# Code Quality & Formatting
black>=23.0.0               # Code formatter
flake8>=6.0.0               # Linting and style checking
mypy>=1.0.0                 # Static type checking
isort>=5.12.0               # Import sorting
bandit>=1.7.0               # Security linting

# Git Hooks & Automation
pre-commit>=3.0.0           # Git pre-commit hooks
commitizen>=3.0.0           # Conventional commits

# Enhanced Testing
pytest>=7.0.0               # Advanced testing framework
pytest-flask>=1.3.0         # Flask-specific pytest utilities
pytest-cov>=4.0.0           # Coverage reporting
pytest-mock>=3.10.0         # Mocking utilities
factory-boy>=3.3.0          # Test data generation

# Database & ORM
Flask-SQLAlchemy>=3.0.0     # SQLAlchemy ORM integration
Flask-Migrate>=4.0.0        # Database migrations
alembic>=1.12.0             # Database versioning

# Real-time & Performance
Flask-SocketIO>=5.3.0       # WebSocket support for real-time features
redis>=4.6.0                # Caching and session storage
Flask-Caching>=2.1.0        # Caching utilities

# Security & Monitoring
Flask-Talisman>=1.1.0       # Security headers
Flask-Limiter>=3.5.0        # Rate limiting
sentry-sdk[flask]>=1.32.0   # Error monitoring

# Development Server
Werkzeug>=3.0.0             # WSGI utilities with reloader

# Documentation
Sphinx>=7.0.0               # Documentation generation
sphinx-rtd-theme>=1.3.0     # Read the Docs theme

# API Development
Flask-CORS>=4.0.0           # Cross-origin resource sharing
marshmallow>=3.20.0         # Serialization/deserialization
webargs>=8.3.0              # Request parsing

# Debugging & Profiling
flask-debugtoolbar>=0.13.0  # Debug toolbar
memory-profiler>=0.61.0     # Memory usage profiling
line-profiler>=4.1.0        # Line-by-line profiling

# File Watching (for auto-reload)
watchdog>=3.0.0             # File system monitoring